% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DomesticCustomCredit.R
\name{DomesticCustomCredit_customer_erp_delete}
\alias{DomesticCustomCredit_customer_erp_delete}
\title{清空ERP客户列表}
\usage{
DomesticCustomCredit_customer_erp_delete(erp_token)
}
\value{
无返回值
}
\description{
清空ERP客户列表
}
\examples{
DomesticCustomCredit_customer_erp_delete()
}
