% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DomesticCustomCredit.R
\name{DomesticCustomCredit_MATUREDDEBIT_view}
\alias{DomesticCustomCredit_MATUREDDEBIT_view}
\title{获取ERP债权表数据}
\usage{
DomesticCustomCredit_MATUREDDEBIT_view(erp_token, FDate)
}
\value{
无返回值
}
\description{
获取ERP债权表数据
}
\examples{
DomesticCustomCredit_MATUREDDEBIT_view()
}
