% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DomesticCustomCredit.R
\name{DomesticCustomCredit_Receive_insert}
\alias{DomesticCustomCredit_Receive_insert}
\title{ERP收款数据插入DMS正式表}
\usage{
DomesticCustomCredit_Receive_insert(dms_token)
}
\value{
无返回值
}
\description{
ERP收款数据插入DMS正式表
}
\examples{
DomesticCustomCredit_Receive_insert()
}
