% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DomesticCustomCredit.R
\name{DomesticCustomCredit_Receive_view}
\alias{DomesticCustomCredit_Receive_view}
\title{获取收款和退款数据}
\usage{
DomesticCustomCredit_Receive_view(erp_token, FDate)
}
\value{
无返回值
}
\description{
获取收款和退款数据
}
\examples{
DomesticCustomCredit_Receive_view()
}
