% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrossProfit.R
\name{GrossProfit_archive_view}
\alias{GrossProfit_archive_view}
\title{查询GrossProfit存档}
\usage{
GrossProfit_archive_view(dms_token, FDate)
}
\value{
无返回值
}
\description{
查询GrossProfit存档
}
\examples{
GrossProfit_archive_view()
}
