% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrossProfit.R
\name{GrossProfit_new_view}
\alias{GrossProfit_new_view}
\title{查询当前GrossProfit}
\usage{
GrossProfit_new_view(dms_token, FDate)
}
\value{
无返回值
}
\description{
查询当前GrossProfit
}
\examples{
GrossProfit_new_view()
}
