% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefundOverduePayments.R
\name{RefundOverduePayments_archive_view}
\alias{RefundOverduePayments_archive_view}
\title{查询RefundOverduePayments存档}
\usage{
RefundOverduePayments_archive_view(dms_token, FDate)
}
\value{
无返回值
}
\description{
查询RefundOverduePayments存档
}
\examples{
RefundOverduePayments_archive_view()
}
