% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefundOverduePayments.R
\name{RefundOverduePayments_new_view}
\alias{RefundOverduePayments_new_view}
\title{查询当前RefundOverduePayments}
\usage{
RefundOverduePayments_new_view(dms_token)
}
\value{
无返回值
}
\description{
查询当前RefundOverduePayments
}
\examples{
RefundOverduePayments_new_view()
}
