% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefundOverduePayments.R
\name{RefundOverduePayments_upload}
\alias{RefundOverduePayments_upload}
\title{上传逾期表}
\usage{
RefundOverduePayments_upload(dms_token, FDate)
}
\value{
无返回值
}
\description{
上传逾期表
}
\examples{
RefundOverduePayments_upload()
}
